Add nonwalkable volume generator
Fix unreachable islands inside of solid objects.

Unity's navmesh generation only looks at intersection of colliders and
doesn't consider convex objects to be solid -- so it creates islands
inside of them. Apparently this is an age-old recast problem.

There's a solution to clean these, but it doesn't work with nav mesh
links:
https://assetstore.unity.com/packages/tools/ai/navmesh-cleaner-151501

It does some fancy stuff with the navmesh triangles and doesn't seem to
use navigability to determine where to place its meshes, so :shrug:.